new: AssignControlPlaneToContext
- ref: add comments to funcs in kubeconfig
- ref: remove useless elements
